Rename iOS namespace to Apple
